Split lsp/handlers.rs into per-request handler files
Splits handlers.rs from 1375 to 223 lines by extracting 4 handler
modules:

- hover.rs (280 lines): hover handler + enrich_hover_detail
- completions.rs (553 lines): completion handler + variant/JSX helpers
- goto_def.rs (141 lines): go-to-definition + import path resolution
- code_actions.rs (224 lines): code actions + import quick-fix helpers

handlers.rs retains: trait impl delegation, references, symbols,
formatting handlers.

Purely structural - no behavioral changes. All 227 LSP integration
tests pass.
